+++ gradle-plugins/react/brownfield/src/main/kotlin/com/callstack/react/brownfield/processors/ProguardProcessor.kt
@@ -1,61 +1,24 @@
 package com.callstack.react.brownfield.processors
 
-import com.callstack.react.brownfield.exceptions.TaskNotFound
-import com.callstack.react.brownfield.shared.Logging
+import com.callstack.react.brownfield.shared.ExplodeAarTask
 import com.callstack.react.brownfield.utils.Utils
 import org.gradle.api.Project
-import org.gradle.api.file.RegularFileProperty
 import org.gradle.api.tasks.TaskProvider
 import java.io.File
 
 class ProguardProcessor(val project: Project) {
-    fun processConsumerFiles(
+    fun processFiles(
         proguardRules: List<File>,
         capitalizedVariantName: String,
+        explodeTask: TaskProvider<ExplodeAarTask>,
     ) {
         val mergeTaskName = "merge${capitalizedVariantName}ConsumerProguardFiles"
-        val mergeFileTask = project.tasks.named(mergeTaskName)
-
-        if (!mergeFileTask.isPresent) {
-            throw TaskNotFound("Task $mergeTaskName not found")
-        }
-
-        mergeFileTask.get().doLast {
-            val outputFile = it.outputs.files.singleFile
-            doLast(proguardRules, outputFile)
-        }
-    }
-
-    fun processGeneratedFiles(
-        proguardRules: List<File>,
-        capitalizedVariantName: String,
-    ) {
-        val mergeGenerateProguardTask: TaskProvider<*>?
-        val mergeName = "merge${capitalizedVariantName}GeneratedProguardFiles"
-        mergeGenerateProguardTask = project.tasks.named(mergeName)
-
-        mergeGenerateProguardTask.get().doLast {
-            val outputFile = it.outputs.files.singleFile
-            doLast(proguardRules, outputFile)
-        }
-    }
-
-    private fun doLast(
-        files: List<File>,
-        outputFile: File,
-    ) {
-        try {
-            val outputFileToMerge =
-                @Suppress("USELESS_IS_CHECK")
-                if (outputFile is File) {
-                    outputFile
-                } else {
-                    (outputFile as? RegularFileProperty)?.get()?.asFile
-                        ?: error("Invalid output file")
-                }
-            Utils.mergeFiles(files, outputFileToMerge)
-        } catch (e: NoSuchFileException) {
-            Logging.log(e)
+        project.tasks.matching { it.name == mergeTaskName }.configureEach { task ->
+            task.dependsOn(explodeTask)
+            task.doLast {
+                val outputFile = task.outputs.files.singleFile
+                Utils.mergeFiles(proguardRules, outputFile)
+            }
         }
     }
 }