+++ gradle-plugins/react/brownfield/src/main/kotlin/com/callstack/react/brownfield/processors/VariantTaskProvider.kt
@@ -1,7 +1,6 @@
 package com.callstack.react.brownfield.processors
 
-import com.android.build.gradle.internal.tasks.factory.dependsOn
-import com.callstack.react.brownfield.exceptions.TaskNotFound
+import com.android.build.api.variant.LibraryVariant
 import com.callstack.react.brownfield.shared.ExplodeAarTask
 import com.callstack.react.brownfield.shared.MergeClassesTask
 import com.callstack.react.brownfield.utils.AndroidArchiveLibrary
@@ -13,7 +12,6 @@ import org.gradle.api.Project
 import org.gradle.api.Task
 import org.gradle.api.UnknownTaskException
 import org.gradle.api.tasks.TaskProvider
-import java.io.File
 
 class VariantTaskProvider(val project: Project) {
     fun bundleTaskProvider(
@@ -29,77 +27,28 @@ class VariantTaskProvider(val project: Project) {
         }
     }
 
-    fun processDataBinding(
-        bundleTask: TaskProvider<Task>,
-        aarLibraries: Collection<AndroidArchiveLibrary>,
-        variantName: String,
-    ) {
-        bundleTask.configure { task ->
-            task.doLast {
-                aarLibraries.forEach {
-                    val dataBindingFolder = it.getDataBindingFolder()
-                    if (dataBindingFolder.exists()) {
-                        val filePath = getReBundleFilePath(dataBindingFolder.name, variantName)
-                        File(filePath).mkdirs()
-                        project.copy { copyTask ->
-                            copyTask.from(dataBindingFolder)
-                            copyTask.into(filePath)
-                        }
-                    }
-
-                    val dataBindingLogFolder = it.getDataBindingLogFolder()
-                    if (dataBindingLogFolder.exists()) {
-                        val filePath = getReBundleFilePath(dataBindingLogFolder.name, variantName)
-                        File(filePath).mkdirs()
-                        project.copy { copyTask ->
-                            copyTask.from(dataBindingLogFolder)
-                            copyTask.into(filePath)
-                        }
-                    }
-                }
-            }
-        }
-    }
-
-    private fun getReBundleFilePath(
-        folderName: String,
-        variantName: String,
-    ) = "${DirectoryManager.getReBundleDirectory(
-        variantName,
-    ).path}/$folderName"
-
     fun preBuildTaskByVariant(
-        variantName: String,
-        buildTypeName: String,
-        isVariantDebuggable: Boolean,
+        variant: LibraryVariant,
         explodeAarTask: TaskProvider<ExplodeAarTask>,
     ) {
-        val preBuildTaskPath = "pre${variantName.capitalized()}Build"
-        val preBuildTask = project.tasks.named(preBuildTaskPath)
-
-        if (!preBuildTask.isPresent) {
-            throw TaskNotFound("Can not find $preBuildTaskPath task")
-        }
-
-        preBuildTask.dependsOn(explodeAarTask)
-
+        val variantName = variant.name
         val bundledAssetsVariantName =
             VariantHelper.getBundledAssetsVariantName(
                 variantName = variantName,
-                buildTypeName = buildTypeName,
-                isDebuggable = isVariantDebuggable,
+                buildTypeName = variant.buildType,
+                isDebuggable = variant.debuggable,
             )
         val capitalizedBundledAssetsVariantName = bundledAssetsVariantName.capitalized()
 
         val projectExt = project.extensions.getByType(Extension::class.java)
         val appProject = project.rootProject.project(projectExt.appProjectName)
-        preBuildTask.dependsOn("${appProject.path}:createBundle${capitalizedBundledAssetsVariantName}JsAndAssets")
 
-        if (Utils.isExpoProject(project)) {
+        val jsBundleTaskName = "${appProject.path}:createBundle${capitalizedBundledAssetsVariantName}JsAndAssets"
+
+        variant.lifecycleTasks.registerPreBuild(explodeAarTask, jsBundleTaskName)
+        if (Utils.isExpoProject(project) && Utils.hasExpoUpdates(appProject, variantName)) {
             val updatesResourcesTaskName = VariantHelper.getExpoUpdatesResourcesTaskName(variantName)
-            if (Utils.hasExpoUpdates(appProject, variantName)) {
-                preBuildTask.dependsOn("${appProject.path}:$updatesResourcesTaskName")
-            }
+            variant.lifecycleTasks.registerPreBuild("${appProject.path}:$updatesResourcesTaskName")
         }
     }